Refactor and optimize code across multiple components and services
- Consolidated import statements for better readability in BinAdminPage, CalendarAdminPage, CreateBookingAdminPage, CreateWithdrawAdminPage, and WithdrawAdminPage.
- Simplified state management by removing unnecessary variables and comments.
- Updated fetch functions to directly set state with API responses, eliminating redundant filtering.
- Improved loading state management in LoginPage and WithdrawAdminPage.
- Enhanced API service functions to accept delete status as a parameter for filtering.
- Streamlined async thunk actions in slices for bookings, expenses, products, and roles to handle delete status.
- General code cleanup for improved readability and maintainability.

+++ src/services/roleService.ts
@@ -3,60 +3,43 @@ import { DeleteStatus } from "../model/delete.type";
 import { HttpClient } from "../shared/utils/HttpClient";
 
 export const createRole = async (payload: any): Promise<any> => {
-  const response = await HttpClient.post(
-    `/permissions/role`,
-    payload
-  );
+  const response = await HttpClient.post(`/permissions/role`, payload);
 
   return response;
 };
 
 export const getAllRoles = async (delet = DeleteStatus.ISNOTDELETE): Promise<any> => {
-  const response = await HttpClient.get(
-    `/permissions/role?delete=${delet}`
-  );
+  const response = await HttpClient.get(`/permissions/role?delete=${delet}`);
 
   return response;
 };
 
 export const getAllRolesForPermission = async (): Promise<any> => {
-  const response = await HttpClient.get(
-    `/permissions/role/authz`
-  );
+  const response = await HttpClient.get(`/permissions/role/authz`);
 
   return response;
 };
 
 export const getRoleById = async (roleId: string): Promise<any> => {
-  const response = await HttpClient.get(
-    `/permissions/role/${roleId}`
-  );
+  const response = await HttpClient.get(`/permissions/role/${roleId}`);
 
   return response;
 };
 
 export const updateRoleById = async (body: any): Promise<any> => {
-  const response = await HttpClient.put(
-    `/permissions/role/update/${body.roleId}`,
-    body.data
-  );
+  const response = await HttpClient.put(`/permissions/role/update/${body.roleId}`, body.data);
 
   return response;
 };
 
 export const isDeleteRoleById = async (body: RoleData): Promise<any> => {
-  const response = await HttpClient.post(
-    `/permissions/role/selectDelete/${body._id}`,
-    body
-  );
+  const response = await HttpClient.post(`/permissions/role/selectDelete/${body._id}`, body);
 
   return response;
 };
 
 export const DeleteRoleById = async (roleId: string): Promise<any> => {
-  const response = await HttpClient.delete(
-    `/permissions/role/delete/${roleId}`
-  );
+  const response = await HttpClient.delete(`/permissions/role/delete/${roleId}`);
 
   return response;
 };
